Move `BlockString` to `Language` dir where it belongs
