+++ src/Language/BlockString.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace GraphQL\Utils;
+namespace GraphQL\Language;
 
 use function array_pop;
 use function array_shift;
