+++ src/Language/Lexer.php
@@ -5,7 +5,6 @@ declare(strict_types=1);
 namespace GraphQL\Language;
 
 use GraphQL\Error\SyntaxError;
-use GraphQL\Utils\BlockString;
 use GraphQL\Utils\Utils;
 
 use function chr;
